requirements.txt
asgiref==3.8.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.3
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.6
jwcrypto==1.5.6
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
